requirements.txt
# Core FastAPI and web server dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database and ORM
sqlalchemy>=2.0.23
alembic>=1.13.1

# Redis for caching
redis>=5.0.1

# Data validation and serialization (updated for Python 3.13 compatibility)
pydantic>=2.6.0

# HTTP clients
requests>=2.31.0
httpx>=0.25.2

# Serial communication for ESP32
pyserial>=3.5

# Campbell Scientific CR1000 data logger
pycampbellcr1000>=0.4

# Flask for legacy support
flask>=2.3.3
flask-socketio>=5.3.6

# Additional utilities
python-dotenv>=1.0.0
psutil>=5.9.6

# PDF generation
reportlab>=4.0.0